// src/generated/prisma/models/DailyCalculation.ts
export type DailyCalculationWhereInput = {
  AND?: Prisma.DailyCalculationWhereInput | Prisma.DailyCalculationWhereInput[]
  OR?: Prisma.DailyCalculationWhereInput[]
  NOT?: Prisma.DailyCalculationWhereInput | Prisma.DailyCalculationWhereInput[]
  id?: Prisma.StringFilter<"DailyCalculation"> | string
  periodStart?: Prisma.DateTimeFilter<"DailyCalculation"> | Date | string
  periodEnd?: Prisma.DateTimeFilter<"DailyCalculation"> | Date | string
  tabil?: Prisma.IntFilter<"DailyCalculation"> | number
  cashInHome?: Prisma.IntFilter<"DailyCalculation"> | number
  cashInShop?: Prisma.IntFilter<"DailyCalculation"> | number
  asol?: Prisma.IntFilter<"DailyCalculation"> | number
  sudh?: Prisma.IntFilter<"DailyCalculation"> | number
  deoya?: Prisma.IntFilter<"DailyCalculation"> | number
  personMoneyTotal?: Prisma.IntFilter<"DailyCalculation"> | number
  leftTotal?: Prisma.IntFilter<"DailyCalculation"> | number
  rightTotal?: Prisma.IntFilter<"DailyCalculation"> | number
  difference?: Prisma.IntFilter<"DailyCalculation"> | number
  balanceStatus?: Prisma.EnumDailyCalculationBalanceStatusFilter<"DailyCalculation"> | $Enums.DailyCalculationBalanceStatus
  recordStatus?: Prisma.EnumDailyCalculationRecordStatusFilter<"DailyCalculation"> | $Enums.DailyCalculationRecordStatus
  openedAt?: Prisma.DateTimeFilter<"DailyCalculation"> | Date | string
  closedAt?: Prisma.DateTimeNullableFilter<"DailyCalculation"> | Date | string | null
  createdById?: Prisma.StringFilter<"DailyCalculation"> | string
  createdAt?: Prisma.DateTimeFilter<"DailyCalculation"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"DailyCalculation"> | Date | string
  lastEditedById?: Prisma.StringNullableFilter<"DailyCalculation"> | string | null
  lastEditedAt?: Prisma.DateTimeNullableFilter<"DailyCalculation"> | Date | string | null
  createdBy?: Prisma.XOR<Prisma.UserScalarRelationFilter, Prisma.UserWhereInput>
  lastEditedBy?: Prisma.XOR<Prisma.UserNullableScalarRelationFilter, Prisma.UserWhereInput> | null
  personMoneyEntries?: Prisma.DailyCalculationPersonMoneyListRelationFilter
  mainCalculations?: Prisma.MainCalculationListRelationFilter
}